+++ lib/home/components/categoriesTab.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class CategoriesPage extends StatelessWidget {
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
