+++ lib/model/Category.dart
@@ -1,0 +1,10 @@
+class Category {
+  String name, image;
+  int id;
+
+  Category({
+    this.id,
+    this.name,
+    this.image,
+  });
+}
