--- lib/model/DanhMuc.dart
@@ -1,10 +1,0 @@
-class DanhMuc {
-  String danhMuc, hinhAnhDM;
-  int maDM;
-
-  DanhMuc({
-    this.maDM,
-    this.danhMuc,
-    this.hinhAnhDM,
-  });
-}
