+++ lib/model/History.dart
@@ -1,0 +1,8 @@
+class History {
+  int id, recipeID;
+
+  History({
+    this.id,
+    this.recipeID,
+  });
+}
