--- lib/model/LichSu.dart
@@ -1,9 +1,0 @@
-class LichSu {
-  String maLichSu;
-  int maMon;
-
-  LichSu({
-    this.maLichSu,
-    this.maMon,
-  });
-}
