--- lib/model/MonAn.dart
@@ -1,19 +1,0 @@
-import 'dart:typed_data';
-
-class MonAn {
-  String tenMon, nguyenLieu, cachNau, danhMuc;
-  int maMon, thoiGian;
-  int yeuThich;
-  Uint8List hinhAnh;
-
-  MonAn({
-    this.tenMon,
-    this.nguyenLieu,
-    this.cachNau,
-    this.maMon,
-    this.danhMuc,
-    this.hinhAnh,
-    this.thoiGian,
-    this.yeuThich,
-  });
-}
